Move heartbeat before doing any work.
Scope::join_heartbeat was starting to execute one of the closures
before calling Scope::heartbeat, which meant that only subsequent
recursive calls using the same scope would be able to take
advantage of parallelism.

This patch moves that call before executing any of the closures in
order to give way to parallelism even in case where only two
closures are run from the root.